Moved Node class to its own file
